Build and run tests as C++ programs
Build and run all tests as C++ programs in addition to C.  This helps
get confidence that rseq, when used from a C++ program, behaves well.

+++ include/rseq/rseq.h
@@ -68,7 +68,7 @@ extern unsigned int rseq_flags;
 
 static inline struct rseq *rseq_get_abi(void)
 {
-	return (struct rseq *) (rseq_thread_pointer() + rseq_offset);
+	return (struct rseq *) ((uintptr_t) rseq_thread_pointer() + rseq_offset);
 }
 
 # ifdef __cplusplus
